requirements.txt
torch==2.2.2
torchvision==0.17.2
mmengine==0.10.7
mmcv-lite==2.1.0
mmsegmentation==1.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
Pillow==10.4.0
numpy==1.26.4
tqdm==4.67.1
onnx==1.16.1
onnxruntime==1.18.1
